requirements.txt
streamlit>=1.28.0
openai>=1.0.0
PyPDF2>=3.0.1
python-docx>=0.8.11
langchain>=1.0.0
langchain-community>=0.4.0
langchain-openai>=1.0.0
langchain-text-splitters>=1.0.0
chromadb>=0.4.24